requirements.txt
google-generativeai
sqlalchemy
schedule
requests
python-dotenv
Pillow
fastapi
uvicorn
pydantic
